fix: minor fixes for blog page
