+++ src/app/[locale]/blog/page.jsx
@@ -1,4 +1,4 @@
-import BlogSection from "@/src/components/blog-page-test/BlogSection/BlogSection";
+import BlogSection from "@/src/components/blog-page/BlogSection/BlogSection";
 
 export default function blogPage() {
   return <BlogSection />;
