--- src/components/blog-page-test/BlogArticleSection/BlogArticleSection.module.scss
@@ -1,55 +1,0 @@
-@use "../../../styles/variables" as var;
-@use "../../../styles/mixins" as mx;
-
-.section {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 16px;
-  max-width: 1680px;
-  margin: 85px auto 0 auto;
-  padding: 32px 16px;
-
-  @include mx.media-q("md") {
-    padding: 48px 40px;
-    gap: 24px;
-  }
-
-  @include mx.media-q("xl") {
-    padding: 60px 64px;
-    gap: 60px;
-  }
-
-  @include mx.media-q("xxxl") {
-    padding: 72px 120px;
-    gap: 56px;
-  }
-}
-
-.wrapper {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  padding: 16px 0;
-
-  @include mx.media-q("md") {
-    gap: 24px;
-    justify-content: space-between;
-    padding: 24px 0;
-  }
-  @media only screen and (min-width: 1220px) {
-    flex-direction: row;
-    flex-flow: row-reverse;
-    padding: 0;
-    justify-content: space-between;
-  }
-}
-
-.socIcons {
-  margin-top: 0 !important;
-  justify-content: flex-start !important;
-  @media only screen and (min-width: 1220px) {
-    flex-direction: column;
-  }
-}
