--- src/components/blog-page-test/BlogSection/items.js
@@ -1,42 +1,0 @@
-export const items = [
-  {
-    id: "1",
-    img: "/images/articles/img1.jpg",
-    title: "article_1.title",
-    description: "article_1.text",
-    link: "https://mc.today/uk/pershoprohodets-otrimaye-vsi-vershki-i-vsi-shishki-olga-ivanova-pro-uspihi-skandali-i-plani-baza-trainee/",
-    date: "27.05.2024",
-  },
-  {
-    id: "2",
-    img: "/images/articles/img2.jpg",
-    title: "article_2.title",
-    description: "article_2.text",
-    link: "https://dev.ua/news/lendinh-tvaryn",
-    date: "27.05.2024",
-  },
-  {
-    id: "3",
-    img: "/images/articles/img3.jpg",
-    title: "article_3.title",
-    description: "article_3.text",
-    link: "https://patronbadge.com/ukr/baza_trainee_ukraine/opportunities_it_students",
-    date: "27.05.2024",
-  },
-  {
-    id: "4",
-    img: "/images/articles/img1.jpg",
-    title: "article_1.title",
-    description: "article_4.text",
-    link: "https://mc.today/uk/pershoprohodets-otrimaye-vsi-vershki-i-vsi-shishki-olga-ivanova-pro-uspihi-skandali-i-plani-baza-trainee/",
-    date: "27.05.2024",
-  },
-  {
-    id: "5",
-    img: "/images/articles/img1.jpg",
-    title: "article_1.title",
-    description: "article_4.text",
-    link: "https://mc.today/uk/pershoprohodets-otrimaye-vsi-vershki-i-vsi-shishki-olga-ivanova-pro-uspihi-skandali-i-plani-baza-trainee/",
-    date: "27.05.2024",
-  },
-];
