+++ src/components/blog-page/BlogSection/BlogSection.jsx
@@ -3,62 +3,93 @@
 import { createKey } from "@/src/lib/utils/createKey";
 import BlogCard from "../../shared/BlogCard/BlogCard";
 import InputSearch from "../../shared/inputs/InputSearch/InputSearch";
-import { items } from "./items";
-import styles from "./BlogSection.module.scss";
 import SocialIcons from "../../shared/SocialIcons/SocialIcons";
 import LoadMore from "../../shared/LoadMore/LoadMore";
-import { usePathname, useRouter, useSearchParams } from "next/navigation";
-import { useCallback, useEffect } from "react";
-import { getSearchParamsObject } from "@/src/lib/utils/getSearchParamsObject";
+import { useTranslations } from "next-intl";
+import { useEffect, useState } from "react";
+import { useInfiniteQuery, useQueryClient } from "@tanstack/react-query";
+import stateSorryModal from "@/src/state/stateSorryModal";
+import { getAllBlogArticles } from "@/src/api/blog";
+import { createImageUrl } from "@/src/lib/hooks/createImageUrl";
+import styles from "./BlogSection.module.scss";
+import clsx from "clsx";
+import Loader from "../../shared/loader/Loader";
+import MessageErrorLoading from "../../shared/MessageErrorLoading/MessageErrorLoading";
+import SorryModal from "../../modals/SorryModal/SorryModal";
 
 const BlogSection = () => {
-  const router = useRouter();
-  const pathname = usePathname();
-  const searchParams = useSearchParams();
+  const t = useTranslations("Blog");
+  const [searchQuery, setSearchQuery] = useState("");
 
-  const createQueryString = useCallback(
-    (data) => {
-      const params = new URLSearchParams(searchParams.toString());
-      data?.map((el) => params.set(Object.keys(el), Object.values(el)));
+  const queryClient = useQueryClient();
+  const open = stateSorryModal((state) => state.open);
+  const {
+    data,
+    fetchNextPage,
+    hasNextPage,
+    isFetchingNextPage,
+    isError,
+    isLoading,
+  } = useInfiniteQuery({
+    queryKey: ["blog", searchQuery],
 
-      return params.toString();
+    queryFn: ({ pageParam = 1 }) =>
+      getAllBlogArticles({ page: pageParam, search: searchQuery, limit: 4 }),
+    getNextPageParam: (lastPage, allPages) => {
+      const morePagesExist =
+        lastPage?.pagination.currentPage < lastPage?.pagination.totalPages;
+      return morePagesExist ? allPages.length + 1 : undefined;
     },
-    [searchParams]
-  );
+  });
+
+  const handleSearchChange = (value = "") => {
+    setSearchQuery(value);
+    if (value === "") {
+      queryClient.resetQueries(["blog"], { exact: true });
+    }
+  };
 
   useEffect(() => {
-    const paramsObject = getSearchParamsObject(searchParams);
-    const { page } = paramsObject;
-    if (!page) {
-      router.push(`${pathname}?${createQueryString([{ page: "1" }])}`);
-      return;
+    if (data?.pages[0].results.length === 0) {
+      open();
     }
-  }, []);
-
-  const onSubmit = (value) => {
-    router.push(
-      `${pathname}?${createQueryString([{ page: "1" }, { query: value }])}`
-    );
-  };
-
-  const handlerClickMore = () => {
-    const currentPage = searchParams.get("page");
-    const nextPage = Number(currentPage) + 1;
-    router.push(`${pathname}?${createQueryString([{ page: nextPage }])}`);
-  };
+  }, [data]);
 
   return (
     <section className={styles.section}>
-      <InputSearch onSubmit={onSubmit} />
+      <InputSearch
+        defaultValue={searchQuery}
+        onSubmit={handleSearchChange}
+        placeholder={t("placeholder")}
+      />
+      {isLoading && <Loader />}
+      {isError && <MessageErrorLoading />}
       <div className={styles.wrapper}>
         <SocialIcons classNameCustom={styles.icons} />
-        <ul className={styles.list}>
-          {items.map((el) => (
-            <BlogCard key={createKey()} item={el} pathname={pathname} />
+        <div className={styles.pages}>
+          {data?.pages.map((page) => (
+            <div key={createKey()} className={styles.page}>
+              {page.results.map((item) => (
+                <BlogCard
+                  key={createKey()}
+                  id={item._id}
+                  img={createImageUrl(item.imageUrl)}
+                  title={item.title}
+                  description={item.text}
+                  date={item.date}
+                />
+              ))}
+            </div>
           ))}
-        </ul>
+        </div>
       </div>
-      <LoadMore onClick={handlerClickMore} />
+      <LoadMore
+        className={clsx(styles.loadMore, !hasNextPage && styles.hidden)}
+        disabled={isFetchingNextPage}
+        text={t("load_more")}
+        onClick={() => fetchNextPage()}
+      />
+      <SorryModal handleCallback={handleSearchChange} />
     </section>
   );
 };
