+++ src/components/shared/LoadMore/LoadMore.module.scss
@@ -8,10 +8,16 @@
   padding: 16px 30px;
   margin: 0 auto;
   box-shadow: inset 3px 3px 4px 0 rgba(0, 0, 0, 0.25);
+  transition: all var.$transition-duration-base var.$transition-timing-base 0s;
   &:disabled {
     cursor: not-allowed;
     opacity: 0.7;
     border: 3px solid var.$secondary-gray;
     color: var.$secondary-gray;
+  }
+  @media (any-hover: hover) {
+    &:hover {
+      background-color: var.$primary-hover;
+    }
   }
 }
